[#159] Completed implementation for Vector4
Also fixed a flakey test in ToolsTests

+++ FatedQuest.Libraries/StandardObjects/Position/FVector4I.cpp
@@ -215,7 +215,7 @@ float FVector4I::Magnitude() const
     // Square Root is not exactly performant and we
     // likely do not need an exact value.
     // TODO: [#32] Change Magnitude which does not use Square Root
-    return std::sqrt(m_x * m_x + m_y * m_y + m_z * m_z + m_w * m_w);
+    return static_cast<float>(std::sqrt(m_x * m_x + m_y * m_y + m_z * m_z + m_w * m_w));
 }
 
 int FVector4I::DotProduct(const FVector4I& other) const
